pyproject.toml
[project]
name = "morgen-mcp"
version = "1.0.0"
description = "MCP server that exposes Morgen calendar data as tools"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mcp[cli]",
    "httpx",
    "python-dotenv",
]

[project.scripts]
morgen-mcp = "morgen_diary.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/morgen_diary"]
